# openapi.json
{"openapi": "3.1.0", "info": {"title": "Artwork Proof Agent", "version": "0.1.0"}, "paths": {"/health": {"get": {"summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/ready": {"get": {"summary": "Ready", "operationId": "ready_ready_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/session": {"post": {"summary": "Session", "operationId": "session_session_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionInput"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/uploads": {"post": {"summary": "Upload", "operationId": "upload_uploads_post", "parameters": [{"name": "name", "in": "query", "required": false, "schema": {"type": "string", "default": "artwork", "title": "Name"}}], "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Uploaded"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs": {"get": {"summary": "Runs", "operationId": "runs_runs_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/PublicRun"}, "type": "array", "title": "Response Runs Runs Get"}}}}}}, "post": {"summary": "Start", "operationId": "start_runs_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/StartInput"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicRun"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{ident}": {"get": {"summary": "Get", "operationId": "get_runs__ident__get", "parameters": [{"name": "ident", "in": "path", "required": true, "schema": {"type": "string", "title": "Ident"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicRun"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{ident}/events": {"get": {"summary": "Events", "operationId": "events_runs__ident__events_get", "parameters": [{"name": "ident", "in": "path", "required": true, "schema": {"type": "string", "title": "Ident"}}, {"name": "after", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "After"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/RunEvent"}, "title": "Response Events Runs  Ident  Events Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{ident}/approve": {"post": {"summary": "Accept", "operationId": "accept_runs__ident__approve_post", "parameters": [{"name": "ident", "in": "path", "required": true, "schema": {"type": "string", "title": "Ident"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DecisionInput"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicRun"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{ident}/reject": {"post": {"summary": "Reject", "operationId": "reject_runs__ident__reject_post", "parameters": [{"name": "ident", "in": "path", "required": true, "schema": {"type": "string", "title": "Ident"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DecisionInput"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicRun"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{ident}/clarify": {"post": {"summary": "Clarify", "operationId": "clarify_runs__ident__clarify_post", "parameters": [{"name": "ident", "in": "path", "required": true, "schema": {"type": "string", "title": "Ident"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClarificationInput"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicRun"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{ident}/cancel": {"post": {"summary": "Cancel", "operationId": "cancel_runs__ident__cancel_post", "parameters": [{"name": "ident", "in": "path", "required": true, "schema": {"type": "string", "title": "Ident"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicRun"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{ident}/resume": {"post": {"summary": "Resume", "operationId": "resume_runs__ident__resume_post", "parameters": [{"name": "ident", "in": "path", "required": true, "schema": {"type": "string", "title": "Ident"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicRun"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{ident}/report.json": {"get": {"summary": "Report", "operationId": "report_runs__ident__report_json_get", "parameters": [{"name": "ident", "in": "path", "required": true, "schema": {"type": "string", "title": "Ident"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProofReport"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{ident}/proof.pdf": {"get": {"summary": "Pdf", "operationId": "pdf_runs__ident__proof_pdf_get", "parameters": [{"name": "ident", "in": "path", "required": true, "schema": {"type": "string", "title": "Ident"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"ClarificationInput": {"properties": {"text": {"type": "string", "maxLength": 2000, "title": "Text", "default": ""}, "width": {"anyOf": [{"type": "number", "maximum": 40.0, "exclusiveMinimum": 0.1}, {"type": "null"}], "title": "Width"}, "height": {"anyOf": [{"type": "number", "maximum": 40.0, "exclusiveMinimum": 0.1}, {"type": "null"}], "title": "Height"}}, "type": "object", "title": "ClarificationInput"}, "DecisionInput": {"properties": {"digest": {"type": "string", "maxLength": 64, "minLength": 64, "title": "Digest"}}, "type": "object", "required": ["digest"], "title": "DecisionInput"}, "Finding": {"properties": {"category": {"type": "string", "enum": ["measured", "model-suggested", "requires-human-review"], "title": "Category"}, "severity": {"type": "string", "enum": ["info", "warning", "error"], "title": "Severity"}, "title": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Title"}, "detail": {"type": "string", "maxLength": 800, "minLength": 1, "title": "Detail"}, "evidence_id": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Evidence Id"}}, "additionalProperties": false, "type": "object", "required": ["category", "severity", "title", "detail", "evidence_id"], "title": "Finding"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ProofReport": {"properties": {"summary": {"type": "string", "title": "Summary"}, "findings": {"items": {"$ref": "#/components/schemas/Finding"}, "type": "array", "title": "Findings"}, "specVersion": {"type": "string", "title": "Specversion"}, "artworkHash": {"type": "string", "title": "Artworkhash"}, "version": {"type": "integer", "title": "Version"}}, "type": "object", "required": ["summary", "findings", "specVersion", "artworkHash", "version"], "title": "ProofReport"}, "PublicRun": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "state": {"type": "string", "title": "State"}, "summary": {"type": "string", "title": "Summary"}, "artworkHash": {"type": "string", "title": "Artworkhash"}, "version": {"type": "integer", "title": "Version"}, "width": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Width"}, "height": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Height"}, "previews": {"items": {"type": "string"}, "type": "array", "title": "Previews"}, "events": {"items": {"$ref": "#/components/schemas/RunEvent"}, "type": "array", "title": "Events"}, "report": {"anyOf": [{"$ref": "#/components/schemas/ProofReport"}, {"type": "null"}]}, "reportDigest": {"type": "string", "title": "Reportdigest"}, "receipt": {"anyOf": [{"$ref": "#/components/schemas/Receipt"}, {"type": "null"}]}, "turns": {"type": "integer", "title": "Turns"}, "inputTokens": {"type": "integer", "title": "Inputtokens"}, "outputTokens": {"type": "integer", "title": "Outputtokens"}, "usedMicros": {"type": "integer", "title": "Usedmicros"}, "error": {"type": "string", "title": "Error"}, "model": {"type": "string", "title": "Model"}, "promptVersion": {"type": "string", "title": "Promptversion"}, "created": {"type": "integer", "title": "Created"}, "expires": {"type": "integer", "title": "Expires"}, "approvalExpires": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Approvalexpires"}}, "type": "object", "required": ["id", "name", "state", "summary", "artworkHash", "version", "width", "height", "previews", "events", "report", "reportDigest", "receipt", "turns", "inputTokens", "outputTokens", "usedMicros", "error", "model", "promptVersion", "created", "expires"], "title": "PublicRun"}, "Receipt": {"properties": {"id": {"type": "string", "title": "Id"}, "reportDigest": {"type": "string", "title": "Reportdigest"}, "artworkHash": {"type": "string", "title": "Artworkhash"}, "version": {"type": "integer", "title": "Version"}, "at": {"type": "integer", "title": "At"}, "simulated": {"type": "boolean", "title": "Simulated"}}, "type": "object", "required": ["id", "reportDigest", "artworkHash", "version", "at", "simulated"], "title": "Receipt"}, "RunEvent": {"properties": {"seq": {"type": "integer", "title": "Seq"}, "kind": {"type": "string", "title": "Kind"}, "title": {"type": "string", "title": "Title"}, "detail": {"type": "string", "title": "Detail"}, "at": {"type": "string", "title": "At"}}, "type": "object", "required": ["seq", "kind", "title", "detail", "at"], "title": "RunEvent"}, "SessionInput": {"properties": {"token": {"type": "string", "maxLength": 200, "minLength": 20, "title": "Token"}}, "type": "object", "required": ["token"], "title": "SessionInput"}, "StartInput": {"properties": {"uploadId": {"type": "string", "title": "Uploadid"}, "width": {"anyOf": [{"type": "number", "maximum": 40.0, "exclusiveMinimum": 0.1}, {"type": "null"}], "title": "Width"}, "height": {"anyOf": [{"type": "number", "maximum": 40.0, "exclusiveMinimum": 0.1}, {"type": "null"}], "title": "Height"}}, "type": "object", "required": ["uploadId"], "title": "StartInput"}, "Uploaded": {"properties": {"id": {"type": "string", "title": "Id"}, "hash": {"type": "string", "title": "Hash"}, "name": {"type": "string", "title": "Name"}, "format": {"type": "string", "title": "Format"}, "pages": {"type": "integer", "title": "Pages"}}, "type": "object", "required": ["id", "hash", "name", "format", "pages"], "title": "Uploaded"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}